MasterProfile: Fix hardware tab radius type
Convert the MasterProfile dialog radius value to Float when calling
AppTheme.roundedBg.

Why:
- fix Kotlin compile error in HardwareTabUi
- keep the Master Profiles UI behavior unchanged

+++ app/src/main/java/com/elitedarkkaiser/redmagic/ui/HardwareTabUi.kt
@@ -202,7 +202,7 @@ object HardwareTabUi {
                     panelColor = AppTheme.panelColor,
                     borderColor = AppTheme.borderColor,
                     dp = { value -> deps.dp(value) },
-                    roundedBg = { fill, stroke, radius -> AppTheme.roundedBg(fill, stroke, radius) },
+                    roundedBg = { fill, stroke, radius -> AppTheme.roundedBg(fill, stroke, radius.toFloat()) },
                     actionButton = { text, isDanger, onClick -> deps.actionButton(text, isDanger, onClick) },
                     space = { value -> deps.space(value) },
                     onSave = { name ->
